fix: Code optimization.

+++ seatunnel-connectors/seatunnel-connectors-flink/seatunnel-connector-flink-fake/src/main/java/org/apache/seatunnel/flink/fake/Config.java
@@ -143,7 +143,7 @@ public final class Config {
     /**
      * mock_data_size Default value.
      */
-    public static final long MOCK_DATA_SIZE_DEFAULT_VALUE = 300L;
+    public static final int MOCK_DATA_SIZE_DEFAULT_VALUE = 300;
 
     /**
      * Create data interval, unit is second. Default is 1 second.
